+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- (spec) Added a file export resource type as a resource which packages can provide as part of their deployments and/or feature flags.
+- (cli) Added checking of conflicts between file export resources with `plt check`/`stage check`.
+
 ## 0.7.0-alpha.0 - 2024-04-10
 
 ### Added